if caret was before the first method identifier character, then call graph related context actions did not respond, this is fix and tests
