--- resharper/resharper-unity/test/data/CSharp/Intentions/ContextActions/EnableExpensive/AddComment.cs
@@ -1,19 +1,0 @@
-using System.Diagnostics.CodeAnalysis;
-using UnityEngine;
-
-namespace Performance
-{
-    public class Everything : MonoBehaviour
-    {
-
-        private void Update()
-        {
-            ManyFixed();
-        }
-        
-        private void ManyFixe{caret}d()
-        {
-        }
-
-    }
-}
