--- resharper/resharper-unity/test/data/CSharp/Intentions/ContextActions/EnableExpensive/AddComment.cs.gold
@@ -1,20 +1,0 @@
-﻿using System.Diagnostics.CodeAnalysis;
-using UnityEngine;
-
-namespace Performance
-{
-    public class Everything : MonoBehaviour
-    {
-
-        private void Update()
-        {
-            ManyFixed();
-        }
-        
-        // ReSharper restore Unity.ExpensiveCode
-        private void ManyFixe{caret}d()
-        {
-        }
-
-    }
-}
